@@ -4,7 +4,7 @@ version: 1.0.2
 homepage: https://github.com/contactlutforrahman/flutter_qr_bar_scanner
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0 <3.0.0"
 
 flutter:
