@@ -15,10 +15,10 @@ import 'package:flutter/services.dart';
 
 
 class PreviewDetails {
-  num height;
-  num width;
-  num orientation;
-  int textureId;
+  num? height;
+  num? width;
+  num? orientation;
+  int? textureId;
 
   PreviewDetails(this.height, this.width, this.orientation, this.textureId);
 }
@@ -49,10 +49,10 @@ class FlutterQrReader {
   static QrChannelReader channelReader = new QrChannelReader(_channel);
   //Set target size before starting
   static Future<PreviewDetails> start({
-    @required int height,
-    @required int width,
-    @required QRCodeHandler qrCodeHandler,
-    List<BarcodeFormats> formats = _defaultBarcodeFormats,
+    required int height,
+    required int width,
+    required QRCodeHandler qrCodeHandler,
+    List<BarcodeFormats>? formats = _defaultBarcodeFormats,
   }) async {
     final _formats = formats ?? _defaultBarcodeFormats;
     assert(_formats.length > 0);
@@ -66,31 +66,31 @@ class FlutterQrReader {
     // invokeMethod returns Map<dynamic,...> in dart 2.0
     assert(details is Map<dynamic, dynamic>);
 
-    int textureId = details["textureId"];
-    num orientation = details["surfaceOrientation"];
-    num surfaceHeight = details["surfaceHeight"];
-    num surfaceWidth = details["surfaceWidth"];
+    int? textureId = details["textureId"];
+    num? orientation = details["surfaceOrientation"];
+    num? surfaceHeight = details["surfaceHeight"];
+    num? surfaceWidth = details["surfaceWidth"];
 
     return new PreviewDetails(surfaceHeight, surfaceWidth, orientation, textureId);
   }
 
-  static Future stop() async {
+  static Future stop() {
     channelReader.setQrCodeHandler(null);
-    return await _channel.invokeMethod('stop').catchError(print);
+    return _channel.invokeMethod('stop').catchError(print);
   }
 
   static Future heartbeat() {
     return _channel.invokeMethod('heartbeat').catchError(print);
   }
 
-  static Future<List<List<int>>> getSupportedSizes() {
-    return _channel.invokeMethod('getSupportedSizes').catchError(print);
+  static Future<List<List<int>>?> getSupportedSizes() {
+    return _channel.invokeMethod('getSupportedSizes').catchError(print).then((value) => value as List<List<int>>?);
   }
 }
 
 enum FrameRotation { none, ninetyCC, oneeighty, twoseventyCC }
 
-typedef void QRCodeHandler(String qr);
+typedef void QRCodeHandler(String? qr);
 
 class QrChannelReader {
   QrChannelReader(this.channel) {
@@ -99,7 +99,7 @@ class QrChannelReader {
         case 'qrRead':
           if (qrCodeHandler != null) {
             assert(call.arguments is String);
-            qrCodeHandler(call.arguments);
+            qrCodeHandler!(call.arguments);
           }
           break;
         default:
@@ -109,10 +109,10 @@ class QrChannelReader {
     });
   }
 
-  void setQrCodeHandler(QRCodeHandler qrch) {
+  void setQrCodeHandler(QRCodeHandler? qrch) {
     this.qrCodeHandler = qrch;
   }
 
   MethodChannel channel;
-  QRCodeHandler qrCodeHandler;
+  QRCodeHandler? qrCodeHandler;
 }
