@@ -4,80 +4,80 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_qr_bar_scanner/flutter_qr_bar_scanner.dart';
 import 'package:native_device_orientation/native_device_orientation.dart';
 
 final WidgetBuilder _defaultNotStartedBuilder = (context) => new Text("Loading Scanner Camera...");
 final WidgetBuilder _defaultOffscreenBuilder = (context) => new Text("Scanner Camera Paused.");
-final ErrorCallback _defaultOnError = (BuildContext context, Object error) {
+final ErrorCallback _defaultOnError = (BuildContext context, Object? error) {
   print("Error reading from scanner camera: $error");
   return new Text("Error reading from scanner camera...");
 };
 
-typedef Widget ErrorCallback(BuildContext context, Object error);
+typedef Widget ErrorCallback(BuildContext context, Object? error);
 
 class QRBarScannerCamera extends StatefulWidget {
   QRBarScannerCamera({
-    Key key,
-    @required this.qrCodeCallback,
+    Key? key,
+    required this.qrCodeCallback,
     this.child,
     this.fit = BoxFit.cover,
-    WidgetBuilder notStartedBuilder,
-    WidgetBuilder offscreenBuilder,
-    ErrorCallback onError,
+    WidgetBuilder? notStartedBuilder,
+    WidgetBuilder? offscreenBuilder,
+    WidgetBuilder? onPermissionDeclined,
+    ErrorCallback? onError,
     this.formats,
   })  : notStartedBuilder = notStartedBuilder ?? _defaultNotStartedBuilder,
         offscreenBuilder = offscreenBuilder ?? notStartedBuilder ?? _defaultOffscreenBuilder,
+        onPermissionDeclined = onPermissionDeclined ?? notStartedBuilder ?? _defaultOffscreenBuilder,
         onError = onError ?? _defaultOnError,
         assert(fit != null),
         super(key: key);
 
   final BoxFit fit;
-  final ValueChanged<String> qrCodeCallback;
-  final Widget child;
+  final ValueChanged<String?> qrCodeCallback;
+  final Widget? child;
   final WidgetBuilder notStartedBuilder;
   final WidgetBuilder offscreenBuilder;
+  final WidgetBuilder onPermissionDeclined;
   final ErrorCallback onError;
-  final List<BarcodeFormats> formats;
+  final List<BarcodeFormats>? formats;
 
   @override
   QRBarScannerCameraState createState() => new QRBarScannerCameraState();
 }
 
 class QRBarScannerCameraState extends State<QRBarScannerCamera> with WidgetsBindingObserver {
-  var shouldResumeOnAppLifecycleChanged = true;
-
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
   }
 
   @override
   dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     super.dispose();
   }
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
-    if (shouldResumeOnAppLifecycleChanged) {
-      if (state == AppLifecycleState.resumed) {
-        setState(() => onScreen = true);
-      } else {
-        if (_asyncInitOnce != null && onScreen) {
-          FlutterQrReader.stop();
-        }
-        setState(() {
-          onScreen = false;
-          _asyncInitOnce = null;
-        });
+    if (state == AppLifecycleState.resumed) {
+      setState(() => onScreen = true);
+    } else {
+      if (_asyncInitOnce != null && onScreen) {
+        FlutterQrReader.stop();
       }
+      setState(() {
+        onScreen = false;
+        _asyncInitOnce = null;
+      });
     }
   }
 
   bool onScreen = true;
-  Future<PreviewDetails> _asyncInitOnce;
+  Future<PreviewDetails>? _asyncInitOnce;
 
   Future<PreviewDetails> _asyncInit(num height, num width) async {
     var previewDetails = await FlutterQrReader.start(
@@ -87,12 +87,6 @@ class QRBarScannerCameraState extends State<QRBarScannerCamera> with WidgetsBind
       formats: widget.formats,
     );
     return previewDetails;
-  }
-
-  void startScanning() {
-    setState(() {
-      _asyncInitOnce = null;
-    });
   }
 
   /// This method can be used to restart scanning
@@ -108,11 +102,10 @@ class QRBarScannerCameraState extends State<QRBarScannerCamera> with WidgetsBind
 
   /// This method can be used to manually stop the
   /// camera.
-  Future stop() async {
-    return await FlutterQrReader.stop();
-    /*(() async {
+  void stop() {
+    (() async {
       await FlutterQrReader.stop();
-    })();*/
+    })();
   }
 
   @override
@@ -140,13 +133,16 @@ class QRBarScannerCameraState extends State<QRBarScannerCamera> with WidgetsBind
             case ConnectionState.done:
               if (details.hasError) {
                 debugPrint(details.error.toString());
+                if (details.error is PlatformException) {
+                  return widget.onPermissionDeclined(context);
+                }
                 return widget.onError(context, details.error);
               }
               Widget preview = new SizedBox(
                 height: constraints.maxHeight,
                 width: constraints.maxWidth,
                 child: Preview(
-                  previewDetails: details.data,
+                  previewDetails: details.data!,
                   targetHeight: constraints.maxHeight,
                   targetWidth: constraints.maxWidth,
                   fit: widget.fit,
@@ -157,7 +153,7 @@ class QRBarScannerCameraState extends State<QRBarScannerCamera> with WidgetsBind
                 return new Stack(
                   children: [
                     preview,
-                    widget.child,
+                    widget.child!,
                   ],
                 );
               }
@@ -176,20 +172,20 @@ class Preview extends StatelessWidget {
   final double height;
   final double width;
   final double targetWidth, targetHeight;
-  final int textureId;
-  final int orientation;
+  final int? textureId;
+  final int? orientation;
   final BoxFit fit;
 
   Preview({
-    @required PreviewDetails previewDetails,
-    @required this.targetHeight,
-    @required this.targetWidth,
-    @required this.fit,
+    required PreviewDetails previewDetails,
+    required this.targetHeight,
+    required this.targetWidth,
+    required this.fit,
   })  : assert(previewDetails != null),
         textureId = previewDetails.textureId,
-        height = previewDetails.height.toDouble(),
-        width = previewDetails.width.toDouble(),
-        orientation = previewDetails.orientation;
+        height = previewDetails.height!.toDouble(),
+        width = previewDetails.width!.toDouble(),
+        orientation = previewDetails.orientation as int?;
 
   @override
   Widget build(BuildContext context) {
@@ -201,7 +197,7 @@ class Preview extends StatelessWidget {
 
         int baseOrientation = 0;
         if (orientation != 0 && (width > height)) {
-          baseOrientation = orientation ~/ 90;
+          baseOrientation = orientation! ~/ 90;
           frameHeight = height;
           frameWidth = width;
         } else {
@@ -209,7 +205,7 @@ class Preview extends StatelessWidget {
           frameHeight = width;
         }
 
-        int nativeOrientationInt;
+        late int nativeOrientationInt;
         switch (nativeOrientation) {
           case NativeDeviceOrientation.landscapeLeft:
             nativeOrientationInt = Platform.isAndroid ? 3 : 1;
@@ -232,7 +228,7 @@ class Preview extends StatelessWidget {
             child: new SizedBox(
               height: frameHeight,
               width: frameWidth,
-              child: new Texture(textureId: textureId),
+              child: new Texture(textureId: textureId!),
             ),
           ),
         );
